factory_management/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-Limiter==2.7.0
PyJWT==2.6.0
pymysql==1.0.3